fix: use `null` as default `data` value if action is missing return
In some cases, actions don't need to return anything on the client. With this fix, if an action is
missing a return statement, but the execution succeeded, `data` value will be `null`. This change
enables the execution of `onSuccess` and `onSettled` hook callbacks when action finishes execution
without server/validation errors.

fix #33