feat(validation-errors): support flattening via `flattenValidationErrors` function (#100)
Sometimes it's better to deal with a flattened error object instead of a
formatted one, for instance when you don't need to use nested objects in
validation schemas.

This PR exports a function called `flattenValidationErrors` that does
what it says. Be aware that it works just one level deep, as it discards
nested schema errors. This is a known limitation of this approach, since
it can't prevent key conflicts.

Suppose this is a returned formatted `validationErrors` object: 

```typescript
validationErrors = {
  _errors: ["Global error"],
  username: {
    _errors: ["Too short", "Username is invalid"],
  },
  email: {
    _errors: ["Email is invalid"],
  }
}
```

After passing it to `flattenValidationErrors`:

```typescript
import { flattenValidationErrors } from "next-safe-action";

const flattenedErrors = flattenValidationErrors(validationErrors);
```

It becomes this:

```typescript
flattenedErrors = {
  rootErrors: ["Global error"],
  fieldErrors: {
    username: ["Too short", "Username is invalid"],
    email: ["Email is invalid"],
  }
}
```