refactor(hooks): use separate type for server state in `useOptimisticAction` (#134)
This PR improves the design of the `useOptimisticAction` hook. Previously, the `optimisticData` had to be of the same type of action's return type. This isn't great, because the only job of a Server Action in a optimistic workflow is to mutate the data on the server. The actual state (and type) that matters is the one coming from the parent Server Component, so `updateFn`'s `prevState` should match that type. The actual result is still available in hooks callbacks and `result` property returned from the hook, though.

re #127