feat(hooks): return `executeAsync` from `useAction` and `useOptimisticAction` hooks (#147)
This PR adds the `executeAsync` function to the return object of `useAction` and `useOptimisticAction` hooks. It's currently not possible to add it to the `useStateAction` hook. As the name suggests, `executeAsync` returns a Promise with the same result of the `safeActionFn` you pass to the hook, and it allows to await the result of the action.

re #137, #72, #94