feat: add action execution callbacks (#161)
This PR adds support for providing optional action execution callbacks when defining an action, invoked after the action succeeds/fails, on the server side. The callbacks are: `onSuccess`, `onError` and `onSettled`, the same as the hook ones, minus `onExecute` (since it provides the same functionality as middleware functions in this context).