refactor: remove additional `fetchError` prop from hook return object
This commit removes the `fetchError` prop from hook return object. Now the returned object is of
type `SafeActionResult`. If an error gets thrown inside an action, and the `handleServerError` init
option rethrows it, now the `execute`/`executeAsync` functions throw it inside the component, so the
handling is up to the user.