refactor: remove additional `fetchError` prop from hook return object (#262)
Code in this PR removes the `fetchError` prop from hook return object.
Now the returned object is of type `SafeActionResult`. This means that
if an error gets thrown inside an action, and the `handleServerError`
init option rethrows it instead of setting the `serverError` prop, now
the `execute`/`executeAsync` functions will throw it too inside the
component, so the handling is managed by the user. This is the same
behavior as directly calling a throwable safe action inside the
component.