refactor: rename `schema` method to `inputSchema`, and deprecate it (#339)
The code in this PR renames the `schema` instance method to
`inputSchema`.
Since introducing output validation with the `outputSchema` method, I
have considered renaming `schema` to `inputSchema` to make its purpose
clearer.

Note that the `schema` method will remain as an alias for `inputSchema`,
but will be removed in a future release.